pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "html-to-markdown>=1.9.0",
    "langchain>=1.0.8",
    "langchain-mistralai>=1.0.1",
    "langchain-ollama>=1.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "werkzeug>=3.1.3",
    "youtube-transcript-api>=1.2.2",
]
